refactor assessment, landing page and user to pages directory
