+++ dottie-app/src/pages/account/password.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { PasswordForm } from '../../components/user/PasswordUpdateForm';
+import { PasswordForm } from '../../pages/user/PasswordUpdateForm';
 
 export default function PasswordPage() {
   // In production, this would come from auth context
