+++ dottie-app/src/pages/assessment/__tests__/DevelopingPattern.test.tsx
