+++ dottie-app/src/pages/assessment/__tests__/e2e/development/MobileView.spec.ts
