+++ dottie-app/src/pages/assessment/period-duration/__tests__/PeriodDuration.test.tsx
