+++ dottie-app/src/pages/assessment/symptoms/__tests__/Symptoms.test.tsx
