+++ dottie-app/src/pages/assessment/symptoms/page.tsx
